backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
python-multipart==0.0.9
httpx==0.27.2
transformers==4.44.2
torch>=2.2.0
sentence-transformers==3.0.1
numpy==1.26.4
opencv-python-headless==4.10.0.84
Pillow==10.4.0
